requirements.txt
openai>=1.2.3
pydantic>=2.0
python-dotenv>=1.0.0
pytesseract>=0.3.10
Pillow>=9.0.0
PyMuPDF>=1.22.0
filetype>=1.2.0
python-magic>=0.4.27
